refactor: extract camera_ready leaf helpers to _util.py (#3385)
First slice of the camera_ready_campaign.py decomposition. Moves pure leaf
helpers (hashers, JSON converters, sanitizers, path/time utilities, kinematics
normalizer) from camera_ready_campaign.py into a new
robot_sf/benchmark/camera_ready/_util.py module.

Changes:
- New robot_sf/benchmark/camera_ready/ package with _util.py containing
  _repo_relative, _utc_now, _hash_payload, _sha256_payload, _sha256_file,
  _jsonable, _jsonable_repo_relative, _sanitize_csv_cell,
  _normalized_kinematics_matrix.
- camera_ready_campaign.py: import from _util and re-export (# noqa: F401)
  for backward compatibility. Remove ~99 lines of inline definitions.
  Drop unused hashlib and UTC imports.

No behavior change. All 112 tests in test_camera_ready_campaign.py pass.
